Fix mapelites history buffer
