+++ elm/elm_main.py
@@ -3,70 +3,29 @@ from omegaconf import OmegaConf
 
 from elm.constants import SRC_PATH
 from elm.diff_model import DiffModel
-from elm.environments import Sodarace
+from elm.environments import IMAGE_SEED, ImageOptim, Sodarace
+from elm.environments.sodaracer import SQUARE_SEED
 from elm.map_elites import MAPElites
-
-TEST_SEED = {
-    "program_str": """from elm.environments.sodaracer.walker import walker_creator
-
-
-def make_square(wc, x0, y0, x1, y1):
-    \"\"\"Make a square with top left x0,y0 and top right x1,y1.\"\"\"
-    j0 = wc.add_joint(x0, y0)
-    j1 = wc.add_joint(x0, y1)
-    j2 = wc.add_joint(x1, y1)
-    j3 = wc.add_joint(x1, y0)
-    return j0, j1, j2, j3
-
-
-def make_walker():
-    wc = walker_creator()
-
-    # the main body is a square
-    sides = make_square(wc, 0, 0, 10, 10)
-    center = wc.add_joint(5, 5)
-
-    # connect the square with distance muscles
-    for k in range(len(sides) - 1):
-        wc.add_muscle(sides[k], sides[k + 1])
-    wc.add_muscle(sides[3], sides[0])
-
-    # one prong of the square is a distance muscle
-    wc.add_muscle(sides[3], center)
-
-    # the other prongs from the center of the square are active
-    wc.add_muscle(sides[0], center, False, 5.0, 0.0)
-    wc.add_muscle(sides[1], center, False, 10.0, 0.0)
-    wc.add_muscle(sides[2], center, False, 2.0, 0.0)
-
-    return wc.get_walker()
-""",
-    "result_dict": {
-        "joints": [(0, 0), (0, 10), (10, 10), (10, 0), (5, 5)],
-        "muscles": [
-            [0, 1, {"type": "distance", "amplitude": 0.0, "phase": 0.0}],
-            [1, 2, {"type": "distance", "amplitude": 0.0, "phase": 0.0}],
-            [2, 3, {"type": "distance", "amplitude": 0.0, "phase": 0.0}],
-            [3, 0, {"type": "distance", "amplitude": 0.0, "phase": 0.0}],
-            [3, 4, {"type": "distance", "amplitude": 0.0, "phase": 0.0}],
-            [0, 4, {"type": "muscle", "amplitude": 5.0, "phase": 0.0}],
-            [1, 4, {"type": "muscle", "amplitude": 10.0, "phase": 0.0}],
-            [2, 4, {"type": "muscle", "amplitude": 2.0, "phase": 0.0}],
-        ],
-    },
-}
 
 
 class ELM:
     def __init__(self, cfg) -> None:
         self.cfg = cfg
         self.diff_model = DiffModel(self.cfg)
-        self.seed = TEST_SEED
-        self.environment = Sodarace(
-            seed=TEST_SEED,
-            diff_model=self.diff_model,
-            eval_steps=self.cfg.evaluation_steps,
-        )
+        if self.cfg.env_name == "sodarace":
+            self.seed = SQUARE_SEED
+            self.environment = Sodarace(
+                seed=self.seed,
+                diff_model=self.diff_model,
+                eval_steps=self.cfg.evaluation_steps,
+            )
+        elif self.cfg.env_name == "imageoptim":
+            self.seed = IMAGE_SEED
+            self.environment = ImageOptim(
+                seed=self.seed,
+                diff_model=self.diff_model,
+                eval_steps=self.cfg.evaluation_steps,
+            )
         self.map_elites = MAPElites(
             self.environment,
             n_bins=self.cfg.behavior_n_bins,
